Revert "Fixed login system"
