+++ ios/Flutter/Release.xcconfig
@@ -1,2 +1,1 @@
-#include "Pods/Target Support Files/Pods-Runner/Pods-Runner.release.xcconfig"
 #include "Generated.xcconfig"
