--- lib/utils/helper.dart
@@ -1,43 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:junkjar/pages/FirstScreen.dart';
-import 'package:junkjar/pages/Home.dart';
-import 'package:shared_preferences/shared_preferences.dart';
-
-import 'Colours.dart';
-
-class Helper extends StatefulWidget {
-  @override
-  _HelperState createState() => _HelperState();
-}
-
-class _HelperState extends State<Helper> {
-  void firstPageChecker() async {
-    SharedPreferences prefs = await SharedPreferences.getInstance();
-    bool isFirstLaunch = prefs.getBool('isFirstLaunch') ?? true;
-
-      if (isFirstLaunch) {
-        Navigator.pushAndRemoveUntil(
-            context,
-            MaterialPageRoute(builder: (context) => MyHomePage(prefs: prefs)),
-            (Route<dynamic> route) => false);
-        // very first launch since install
-      } else {
-        Navigator.pushAndRemoveUntil(
-            context,
-            MaterialPageRoute(builder: (context) => FirstScreen(prefs: prefs)),
-            (Route<dynamic> route) => false);
-      }
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    firstPageChecker();
-  }
-
-  Widget build(BuildContext context) {
-    return Container(
-      color: MyColors.black,
-    );
-  }
-}
