+++ pubspec.yaml
@@ -21,15 +21,12 @@ dependencies:
     sdk: flutter
 
   cupertino_icons: ^0.1.2
-  cloud_firestore: ^0.13.2+1
   gradient_widgets: ^0.5.2
   google_fonts: ^0.3.7
   flare_flutter: ^2.0.1
   firebase_auth: ^0.15.4
   google_sign_in: ^4.1.1
   flutter_signin_button: ^1.0.0
-  shared_preferences: ^0.5.6+1
-  uuid: ^2.0.4
 
 dev_dependencies:
   flutter_test:
